Change Map value Integer to Long in Expression package

+++ src/main/java/Grammar/Expression/BinaryArithExpr.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public record BinaryArithExpr(Expression left, String op, Expression right) implements Expression {
     @Override
-    public long eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
+    public long eval(Map<String,Long> bindings, Player p, Territory t) throws EvalError {
         long lv = left.eval(bindings,p,t);
         long rv = right.eval(bindings,p,t);
         switch (op) {
