+++ src/test/java/Grammar/Plan/RegionCommandTest.java
@@ -15,7 +15,7 @@ public class RegionCommandTest {
     public void investCommand(){
         RegionCommand regionCommand = new RegionCommand("invest", new Expression() {
             @Override
-            public long eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
+            public long eval(Map<String, Long> bindings, Player p, Territory t) throws EvalError {
                 return 50;
             }
             @Override
@@ -33,7 +33,7 @@ public class RegionCommandTest {
     public void collectCommand(){
         RegionCommand regionCommand = new RegionCommand("collect", new Expression() {
             @Override
-            public long eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
+            public long eval(Map<String, Long> bindings, Player p, Territory t) throws EvalError {
                 return 10;
             }
             @Override
